Add Kafka Serde implementations for Kafka Streams applications
 - This commit adds three implementations of Kafka's Serde interface
which provides a wrapper for a Serializer and Deserializer. The
three implementations wrap the AvroKafkaSerializer/Deserializer, the
JsonSchemaKafkaSerializer/Deserializer and the
ProtobufKafkaSerializer/Deserializer. See issue #641 for more info.
 - It also adds a workaround for the issue where Kafka Deserializers
throw a `java.lang.IllegalStateException: Method not supported on an
outbound message` error (see #702 and #808) due to the use of
JAX-RS classes in the CompatibleClient class.

Signed-off-by: Andrew Borley <borley@uk.ibm.com>