feat: resolve artifacts by their content hash
This addresses Issue #3237, adding the ability to resolve artifacts by their content hash in the Kafka serde. Note that resolving a schema by content hash directly from their data is not supported, as the schema data is generated in these cases and has a very good chance of not being consistent with the schema content that was actually registered. This would cause a server call and worsen performance.