feat: resolve artifacts by their content hash (#3244)
* feat: resolve artifacts by their content hash

This addresses Issue #3237, adding the ability to resolve artifacts by their content hash in the Kafka serde. Note that resolving a schema by content hash directly from their data is not supported, as the schema data is generated in these cases and has a very good chance of not being consistent with the schema content that was actually registered. This would cause a server call and worsen performance.

* fix: remove print statement and fix javadoc

* refactor: add content hash TTL enforcement test

* docs: add apache license to files

also includes:

* refactor: use getter instead of property directly in ArtifactReferenceImpl

* fix: add missing method to MockRegistryClient

* docs: fix copyright date in new test files

---------

Co-authored-by: Devon Berry <devon.berry@riotgames.com>
Co-authored-by: Carles Arnal <carlesarnal92@gmail.com>