[2.6.x] Update keycloakjs and use access_token instead of id_token for oidc code path (#5371)
* Update keycloakjs version to avoid cyclic redirect

Nonce support in urls was removed in Keycloak 25 to be
fully oidc compatible. But this change is not compatible with
old keycloakjs version. Apicurio used a very old version
and was broken after this keycloak release.

* Use access_token not id_token for Bearer auth

If the oidc client auth library is used (instead of the
default keycloakjs on) the id_token is used as Bearer Token.
This is not standard compatible. ID Tokens only contain user information
and should never be used for authentications.