fix: Wrong groupid registering referenced avro schema
Currently when defining a schema that references another schema, the
apicurio-registry-maven-plugin registers the referenced schema using
the parent groupid and set the referenced schema artifactid to its
fullname (`groupid.artifactid`).
This issue defeats the very purpose of having separate schemas to put some avro
object definition in common.

The proposed solution is to create a `ReferenceArtifactIdentifierExtractor` util that is in charge of parsing the artifact id and group id from the external reference `resource` string.
An AVRO implementation just break down the resource and separate the
artifact id from the group id depending on the position of the last
existing `.` according to AVRO 1.12.0 specification.
Other implementations are just a default one mimicking the current
behavior : groupid = null, artifactid = resource.