Add versionMutabilityEnabled option to Registry CRD for Studio support (#6781)
* Add versionMutabilityEnabled option to Registry CRD for Studio support

This adds a new boolean property 'versionMutabilityEnabled' to the
AppFeaturesSpec in the Registry Custom Resource Definition. When set to
true, this enables the artifact version mutability feature, which allows
artifact versions in DRAFT state to have mutable content. This also unlocks
Studio functionality in the Apicurio Registry UI.

Changes:
- Added APICURIO_REST_MUTABILITY_ARTIFACT_VERSION_CONTENT_ENABLED constant
- Added versionMutabilityEnabled property to AppFeaturesSpec with documentation
- Added operator logic to set environment variable when feature is enabled
- Added integration tests for the new feature

Fixes #6694

* Fix test to verify environment variable value, not just existence

Updated testVersionMutabilityEnabledTrue to check that the environment
variable value is actually "true", not just that the variable exists.
This prevents the test from passing if the value was incorrectly set
to "false" or any other value.