Try to deserialize base64 encoded protobuf schemas.
When checking compatibility rules, the new Protobuf schema is converted from base64-encoded form
into a parsed ProtoFile (these changes were made in PR !6124). However, the same is not done for the
stored schemas. This change adds this conversion to the ProtobufFile class (try to parse as text,
deserialize via base64 and FileDescriptorProto if this fails). This fixes rule validation, but might
expose bugs w.r.t nested type names as reported in issue #6832.