Implemented Protobuf type name normalization for nested types.
The compatibility checker now uses the parent type to try and normalize nested type definitions. It
splits the parent type's name to get a list of scopes to look in, and then iterates through those
scopes looking for a defined type for that scope.

Example:

message ParentOne {
  message NestedType {
    uint64 some_value = 1;
  }
  message ParentTwo {
    NestedType nested = 1;
  }
}

In this case, the normalization code would check whether the type `ParentOne.ParentTwo.NestedType`
is defined (it is not), and then check `ParentOne.NestedType`, which exists. This matches the scope
lookup rules of Protocol Buffers more closely than the previous implementation.

It does not change behavior for fully qualified type names at all.